Added SEV pre-processed data
This is the pre-processed data for the Sevilleta LTER site. The study had cover and biomass and both were included here. There were many species sampled for each quadrant inside plots, so the biomass was summed for the species and the mean was taken for each quadrant and treatment for the plots.