Added processed file for MCM
This experiment was conducted in two basins which were congregated here by taking the mean. The data were grouped by block then by treatment. There was no soil organic carbon data for any of the years except 2001, so the abundance data (#/kg of soil) of nematodes and rotifers were used as the carbon indicator.