Added processed file for HBR
This was a litterfall experiment across several different sites. In each site there were a number of stands in which plots were sampled in the fall, spring, and summer. The fall total mass in g/m2 of leaf litter was used and the mean was taken across the plots within a stand.