Added processed file for arctic lter
This was a fertilization experiment that has 4 blocks. Data was separated into species and part of the plant for biomass, so the sum was taken of the parts then the sum was taken for the species. The data for 4 blocks were in separate columns but there was a column for the mean so that was used.