Redid the plum island LTER file
The column for treatment (creek) was removed in the previous processed data for PIE. Just added that back in. Used shoot.mass as the biomass metric. I took the mean of shoot mass per species, then summed the species within a transect, then took the mean for each transect. There is a left and right side of each creek that I suppose should be the actual replicate.